outsource common proxy functionality to base class
