+++ WebVotingProxy/WebVotingProxy.csproj
@@ -7,7 +7,7 @@
     <ProjectGuid>{85F50259-05E7-4BE7-BA0F-6D5B0ADF9A34}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <RootNamespace>WebVotingProxy</RootNamespace>
-    <AssemblyName>TwitchChatVotingProxy</AssemblyName>
+    <AssemblyName>WebVotingProxy</AssemblyName>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
